// pbweberogwcl/src/app/dati-progetto/commons/models/recapiti.ts
/*
 * Copyright Regione Piemonte - 2024 
 * SPDX-License-Identifier: EUPL-1.2 
*/

/* tslint:disable */
export class Recapiti {
  dtFineValidita?: string;
  dtInizioValidita?: string;
  email?: string;
  pec?: string;
  fax?: string;
  flagEmailConfermata?: string;
  flagPecConfermata?: string;
  idRecapiti?: number;
  idUtenteAgg?: number;
  idUtenteIns?: number;
  progrSoggettoProgetto?: number;
  progrSoggettoProgettoSede?: number;
  telefono?: string;
}